+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Commands are explained like so:
 
 ### Fixed
 - Fixed `/modping` command to use moderators roles from the `config.json` file instead of using the role name, removed duplicated code.
+- Fixed quoted messages not being deleted when using the delete button
 
 ## [v2.4.0] 15/05/2022